split code into modules upgrade ts, jest
