+++ package.json
@@ -14,13 +14,13 @@
     "tslib": "^1.10.0"
   },
   "devDependencies": {
-    "@types/jest": "^24.0.15",
+    "@types/jest": "^26.0.20",
     "browserify": "^16.5.1",
-    "jest": "^24.8.0",
+    "jest": "^26.6.3",
     "rimraf": "^2.6.3",
-    "ts-jest": "^24.0.2",
+    "ts-jest": "^26.4.4",
     "tsify": "^4.0.2",
-    "typescript": "^3.5.2"
+    "typescript": "^4.1.3"
   },
   "scripts": {
     "prebuild": "rimraf dist",
