+++ src/builders/agregate.ts
@@ -1,0 +1,26 @@
+import {Aggregate} from './../odata-types';
+
+export function buildAggregate(aggregate: Aggregate | Aggregate[]) {
+    // Wrap single object in an array for simplified processing
+    const aggregateArray = Array.isArray(aggregate) ? aggregate : [aggregate];
+
+    return aggregateArray
+        .map(aggregateItem => {
+            return typeof aggregateItem === "string"
+                ? aggregateItem
+                : Object.keys(aggregateItem).map(aggregateKey => {
+                    const aggregateValue = aggregateItem[aggregateKey];
+
+                    // TODO: Are these always required?  Can/should we default them if so?
+                    if (!aggregateValue.with) {
+                        throw new Error(`'with' property required for '${aggregateKey}'`);
+                    }
+                    if (!aggregateValue.as) {
+                        throw new Error(`'as' property required for '${aggregateKey}'`);
+                    }
+
+                    return `${aggregateKey} with ${aggregateValue.with} as ${aggregateValue.as}`;
+                });
+        })
+        .join(',');
+}
