+++ src/constants.ts
@@ -1,0 +1,2 @@
+export const INDEXOF_REGEX = /(?!indexof)\((\w+)\)/;
+export const ITEM_ROOT = "";
